requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==25.4.0
behave==1.2.6
behave-html-formatter==0.9.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
packaging==26.0
parse==1.21.1
parse_type==0.6.6
pluggy==1.6.0
pycparser==3.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.5
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
